Remove R cropping code
The R cropping code was always a placeholder with reticulate doing the heavy-lifting as the ultimate goal. Now that reticulate is being introduced, it makes sense to cull the R code as it drags down performance.

There are issues getting reticulate to work in the datalabs platform, so I can't validate that the reticulate code is correct, but from what I can tell it should be valid.