Start adding cached shapefiles from Python
To increase performance, we add minimal shapefiles which have precomputed the overlap of administrative zone boundaries and the ratio of area overlap, i.e. the two main unknowns to be integrated with the user-supplied data.

These shapefiles have been generated from a separate python script, for historical reasons. Once the process is perfected, these python scripts may be integrated to this codebase.