Update analyse script + add plot
The analyse script now has two functions, one for each workflow, although the cropping function is a stub. Sourcing the script on shiny startup and calling the functions in the server block seems to work well.

The "non-gridded" workflow is all hardcoded for example purposes for now, but does produce the correct output for the example cleaned data, which was added to relieve development burden for now.