Accept user data for wf2
The upload data button now responds to a file uploaded by the user in workflow 2. Download is still not possible and scottish datazones are the only supported admin zones.

The first column of the uploaded dataset must contain data zone IDs and the last column of the dataset will be plotted. An example format of data is included, wf2-test-data.csv.