Implement upload + download for wf1
The implementation is done using the lessons learned from the upload and download for wf2, with some modifications to account for the lack of native R data structure for raster data. Using the raster library might allow for this in the future but since we already write to/from temporary files I don't see this changing.

Also implements a small change for wf2 - previously, the downloaded data would include the ID twice, once from each dataset. The duplicated ID is now removed, leaving only unique columns from the uploaded data and the data from the relevant shapefile.