Add additional zones within the Cairngorms
Rather than being hardcoded to work with the Cairngorms/data zone
combination, Nuts levels 0-3 are now a valid choice for uploaded data.
Test data is included which is correctly processed - rows should be in
the output if and only if their testval is not 666.

This commit extends two-letter abbreviations for sites and zones, first
absent-mindedly introduced in analyse.py. These have been added to the
shiny interface and are now received and parsed (potentially
inefficiently via if/elif/.../else) by the aggregate function.

Incidentally, commit messages will now conform to the guidlines set by
Tim Pope. RStudio has not wrapped lines to 72 characters in previous
commits. Pope's guidelines can be viewed at:
https://tbaggery.com/2008/04/19/a-note-about-git-commit-messages.html