Reformat Cairngorms to match incoming site data
This commit renames the Cairngorms shapefiles to boundaries.shp, with
the understanding that the meaning of the identical filenames can be
inferred from the directories containing them. This may also set the
scene for automatic, on-the-fly loading of shapefiles.

However, I also noticed that the NUTS regions are available in multiple
CRSes, including EPSG:4326, the given format of the DEIMS shapefiles.
To minimise reprojection inaccuracies, the Cairngorms/NUTSx shapefiles
were recalculated, which unfortunately pollutes the git status.

Finally, the new filenames are updated in analyse.py.