Improve plot output
To map input codes (e.g. 'aa', 'n2') to user-friendly names, a dict is
a natural data structure. Two dicts are added, one for LTSER site names
and one for admin zone names. The friendly names are now included in
the plot title.

The plot output now includes grey shading for zones with missing values
instead of being blank, as well as aligning the legend with the plot.