Load wf2 shapefiles dynamically instead of statically
To test if start up times can be reduced, reimplement the wf2 function
to load composite shapefiles on-the-fly according to input, rather than
selecting a preloaded file.

This works, with no noticeable impact on performance. On testing, it
seems that this has the unexpected side effect of causing Cairngorms
NUTS images to come out at different sizes, although this may have been
introduced in the reprojection of Cairngorms NUTS shapefiles.