Fix packrat configuration and automate installation
A packrat snapshot had not yet been taken, and the development
environment could not be reliably recreated. That is corrected with
this commit. The README has been updated to reflect this, along with
instructions on how to install the R and Python dependencies.

The main app.R script previously depended on the entire tidyverse
library. To reduce the time to recreate the environment, this has been
reduced to only the necessary dplyr and readr dependencies.

Superfluous packrat files have been removed and added to the .gitignore
for convenience.