Parse wf2 data and present columns
This commit only updates the interface to read the input data and
present its columns (minus the first column, which is always taken to
be the ID) in the interface. These are not yet passed to the python
function.

Given the similarity between the introduced output$wf2_columns and
tabular_output reactive elements, it seems likely that these can be
combined eventually. They are left separate for now.