Enable remaining DEIMS + NUTS sites for wf1
Using an updated script, the NUTS metadata has been updated to include
all NUTS regions at all levels. The interface now loads this new
metadata. To avoid errors on startup, defaults have been set for all
NUTS levels. The label for DEIMS selection has been updated to reflect
the availability of the remaining DEIMS sites and the user's selection
is now passed through to python.

To avoid repeated reprojection errors, the global variable declarations
have been removed and replaced by the taking of copies of the reference
datasets. This, along with the inclusion of all NUTS regions now, has
massively reduced the performance of wf1, taking 10-15 seconds to
calculate each time the selection is changed. This is fine for now, as
there are several possible remedies which will be offered to users.

The code lookup dictionaries (e.g. aa > Atelier Alpes) are now global
to allow both workflows to refer to them. Only one is now made use of
in wf1, to look up the name of the DEIMS site passed through from R.
This passthrough is now mandatory, as there is no longer a default in
python.