Enable custom plot title for wf1 + cosmetic changes
In the same way as wf2, on upload of a custom dataset a prompt appears
which is rendered in the title of the plot. This is again implemented
by a new reactive expression, with potential for consolidation in
future. A plot title (custom or default) is now passed to python, which
accepts this and plots against updated template strings.

To match wf2 and reflect the current lack of dynamic data (retrieved
from API), the data selection drop-down list has been removed for wf1
and the prompt changed for both workflows. An irrelevant date filtering
section was also removed for clarity.