Add remaining Scottish national statistical zones
The Cairngorms national park now has shapefiles with subdivisions of:
- intermediate zone (iz)
- health board (hb)
- health integration authority (ia)
- UK parliamentary constituency (up)
- Scottish parliamentary constituency (sp)
- Scottish Council (sc)
- Electoral Ward (ew)
- Travel-to-work areas (tw)

The source files have not been checked-in to the repo but are cited in
a new file in the root Cairngorms directory.

To generate the remaining shapefiles, the generation script has been
updated. First, the geometry of the "debug" output has been set
correctly so that the geodataframes can be saved directly. Second, a
new wrapper function has been added to do the work of opening source
files and saving them somewhere appropriate.