Enable additional Scottish zones
The shapefiles added last commit can now be selected in the interface
and used by the analyse.py function. An example data generator script
and some output have been added for ease of future demos and debugging.

To enable the new shapefiles updates were made in the following places:
shiny/app.r
- UI, data_grouping section - friendly names + codes

python/analyse.py
- preamble, loading of shapefiles - code
- code > friendly name dictionary
- aggregateTabularDataset function - site + zone codes => shapefile map

While this was a relatively quick process, it underscores the need to
further develop a centralised location/script for metadata. This work
can build off the prototype 'hotswapping' branch which makes some of
the needed changes already.