Add excel support for wf2
Since the shiny app passes a data frame/tibble to python, adding excel
support only requires updates to the shiny code. Firstly, the readxl
package has been added to the packrat environment to actually read the
files. All other changes are adjustments to the interface.

To accommodate the fact that multiple sheets can be present in a file,
an option to choose the sheet to read has been added when an excel file
is uploaded. Originally this was just added to the wf2_columns reactive
expression, but then the columns to plot would not update because they
were reliant on output from the same reactive element.

Therefore extra reactive elements were added and the existing elements
tidied so that e.g. they weren't all reading the input file.

Finally, test data is included for verification.