Performance patch for wf1
CRS changes now (again?) mutate global reference shapefiles, so rather
than reprojecting on every user update, reprojections are cached.

This approach drastically increases performance, but introduces a new
issue: as the reference shapefiles are reprojected over and over, they
risk being distorted over the course of an application session where
lots of user data is uploaded.

For most use cases this should be an acceptable risk but ought to be
prevented or quantified in a production-quality system.