Update README to act as overview
The README now acts as an introduction to the repo aimed at high-level
understanding of the problem domain and of the codebase as a whole.

Some README segments have been moved as-is to the relevant documentation
folder. Where this has happened, the placeholder for that directory has
been removed as it is no longer required.