Add introductory tutorial for wf1 only
This is the first attempt at writing an introductory tutorial, following
from updating the main README. The file is only indirectly referenced
there but a more direct link can be added if user feedback suggests this
is necessary.

The tutorial includes screenshots and is designed preferably to be
followed along with by a user, although many will probably be fine just
reading it to get and idea of how workflow 1 works.

Before writing an introduction to workflow 2 in the same style, a quick
user test will be run to check for obvious deficiencies.