Tidy example data
The Nitrous oxide data is formed of two files, which isn't currently
supported for wf1. Since this sends mixed messages, this data is deleted
from the example repo, as alternative sample data exists.

While deleting example data, it seemed prudent to tidy the directory a
little. Now, each file has a purpose:

- `wf1-test-data.tif` allows testing of workflow 1. It would be nice to
upload a dataset which covers all six sites but all datasets tested so
far are larger than I'd like in a git repo.

- `wf2-test-data.csv` allows testing of workflow 2. Three columns of
random integers 1-10 are available for every NUTS region, which
future-proofs nicely for additional sites which are added.

- `fullscot.xlsx` demonstrates Excel support and demonstrates national
statistics.

- `scottish-births-cleaned.csv` is real-world data which is easy to use
in the tutorial.

- `scottish-births-raw.csv` shows the raw births data and demonstrates
the kind of cleaning required for wf2, although this will be documented
explicitly soon.

References to sample data are updated in the README and beginner's
tutorial.