Restructuring: rename sf/deims dirs & add metadata
WARNING: the code as committed here will not work. This is intentional.

This commit begins the rearchitecting of the project - the hardcoded
nature of e.g. the site names, shapefile locations etc. was good to get
off the ground quickly but the time has come to reorganise so that
adding sites (eventually on-the-fly, perhaps) is simple and all code is
modular.

We start that work by renaming the shapefile folders within
shapefiles/deims/ for the DEIMS ID suffix of the site and adding a
metadata.json file to be parsed by machine later. While not yet
finalised, the files as added here should be fairly representative of
the final structure.

To keep commit diffs to a manageable level, we allow commits to break
the code so changes can be introduced gradually in coherent groups.