Restructuring: change sf/zones & add metadata
WARNING: the code as committed here will not work. This is intentional.

We delete the zip containing the full NUTS 2016 data and the extracted
subset of all regions in the EPSG:3857 projection. We also delete two
unused CSVs of generated metadata derived from these.

They are replaced with a proper structure, where within the nuts2016
directory (anticipating adding support for NUTS2021) each level 0-3 has
a directory containing both metadata (for machine-generating composite
shapefiles  and for displaying friendly names) and shapefiles with a
standardised name.

The Scottish Data Zones are similarly restructured, although the other
Scottish national zone sources are missing. This is fine for now as for
the medium-term we only need to dynamically work with NUTS zones.