Restructuring: add Scottish zone metadata
WARNING: the code as committed here will not work. This is
intentional.

To maintain feature parity on completion of this refactor, we need
metadata for all Scottish admin zones for wf2, so we add this here -
correct it in the case of data zones. These names are all taken from the
soon-to-be-deleted vector hardcoded in app.R for the data_grouping
input.

So the correct national zones can be matched to sites we also alter the
deims metadata - nationalZoneDir can now be used to link composite
shapefiles with the zone metadata.

We also restore NUTS_RG_01M_2016_3857.shp.zip - it was deleted to allow
removing NUTS regions from wf1 (which may still happen eventually) -
however it's actually not trivial to excise the relevant code from app.R
so this functionality is best preserved during this transition, with the
option to remove later.