Restructuring: refactor app.R and analyse.py
This is the first attempt at rewriting app.R and analyse.py to utilise
the new metadata directory structure. For circumstantial reasons the
code has only been minimally tested, and so bugs are highly likely.

Due to the size of the commit, the changes are summarised in less detail
than usual.

For app.R:
- deims_LTSER_choices, the vector mapping friendly names to site codes,
is deleted. The input now looks to deims_site_name_mappings, generated
in python, for options
- for wf2 the zone picker with hardcoded choices is deleted and replaced
with reactive input, which looks to deims_site_zone_options

For analyse.py:
- os and json imports are added to parse metadata files on disk
- hardcoded shapefiles (except NUTS for wf1), site names and zone names
are deleted and replaced with code to load these programatically from
the directory structure. Documentation has been started to detail the
process.
- metadata referenced in app.R is generated after the above step
- simplify workflow code to match input to generated metadata. Some
minor reordering has happened to group relevant setup/work code more
logically too.
- rewrite wf1 NUTS code to reproject single zones each time, rather than
the whole GDF. Unless somebody toggles the choice of site an unrealistic
amount this should result in overall better performance