Fix interface issues
There are two issues this commit addresses.

The first is that the list/dict mapping site names to IDs was backwards,
i.e. displaying IDs and passing friendly names. This is easily addressed
by reversing the arguments in the dictionary comprehension.

The second issue is that the reactive input was always empty. After a
lot of experimenting it turned out that it was the subsetting syntax
that was wrong. A simple change from $ to [[]] resolves the problem,
although the options now appear in an ugly order. A note has been made
to resolve this later.

Testing both workflows, the only issue with functionality was that wf2
didn't work at all. The issue was that the "region" variable had been
accidentally copied from wf1. While correcting this, the variables
throughout the function were standardised as "deims_site", "admin_zones"
and "site_name".

Finally, some temporary test code (see testvar) was accidentally left
in. This is removed and reformatted as it ought to have been.