Update shapefile script to work with new structure
The script is updated to handle the entire process of adding a new DEIMS
site to the system, utilising the new metadata directory structure:

- the composite shapefile generating logic is loaded, the only unchanged
element from before
- NUTS shapefiles and metadata are loaded for zones 0-3
- helper functions are defined to provide DEIMS-related functionality
- the function addDeimsSite wraps everything to handle the whole process

addDeimsSite can be called from e.g. an interactive python3 session and
has been successfully tested, although it currently saves DEIMS site
boundaries as zipped shapefiles, which is the default format provided.
This is incompatible with the current site loading code in analyse.py.