Integrate DEIMS site addition to interface
The frontend (wf2 only, for now) now presents a text input and button.
On pasting a DEIMS ID (full or suffix-only) and submitting the
addDeimsSite function from shapefile-generator.py is called with the
user input and the interface is (eventually) updated with the new
option.

The security implications of passing user input to the function have not
been deeply scrutinised yet. The code is currently not public and so
this is unlikely to present any risk.

To facilitate live updates to the interface (re: which sites are
available to choose) another uiOutput/renderUI is added. This references
the reactiveValues deims_site_options to trigger re-renders on updates.
The event handling code for the submit button updates the reactiveValues
directly.

An "interactive" mode is also added to the addDeimsSite function for use
from the interface - passing interactive=True saves generated metadata
which is otherwise discarded and adds it to the global variables which
are passed to R from analyse.py.

Note the use of py$deims_site_name_mappings in app.R - this took forever
to work out but is required. It seems with reticulate that global
variables from python are copied once into the R space and then left.
Trying this code without py$ didn't work, presumably because of the
stale R copy of the dictionary. Whatever the mechanism, the python
updates now work successfully and have been tested.