Update python and R dependencies
Although keeping dependencies up-to-date is just generally a good idea,
the timing of this reflects the fact that, for whatever reason, the code
as it stands doesn't perform well enough to load in datalabs. As
mentioned before, the old dependencies don't work with the newer
versions of python and R now available in datalabs, but the newer
versions introduced here have been successfully tested.