Refactor wf1 to read input from local filesystem
To start integrating the system with wherever it is hosted, we now
handle input locally instead of relying on browser uploads. These are
still allowed but are now added to the local input directory rather than
being processed directly.

Next is to refactor output in a similar way before replicating it all
for wf2.