Refactor wf1 to write output to local filesystem
We complete the wf1 refactor, which now reads and writes data to/from
the local filesystem. Upload and download is still allowed.

We also improve the output button behaviour, as these are now only
rendered when relevant. This was especially noticeable on a clean
installation with no data - the elter logo is now the only output.