Fix and improve installation, upgrade R packages
As I suspected, trying to remove dplyr within packrat seems to have
introduced some inconsistencies, so we refresh packrat.lock completely
with updated packages. While we're at it, we also update
requirements.txt with newer versions.

To accommodate the new directory structure for input and output, we
update the setup.sh script to create these directories for us. Now that
this is done here, we can decouple this from analyse.py.

Finally, we add an additional setup script for extra, datalabs-specific
setup tasks. To embrace DRY it sources the "standard" script.