Refactor/modularise Python code, automate GIS
This is a huge commit, so some details may be lost. Most changes are
derived from 5 files:
        modified:   analyse.py
        modified:   app.R
        new file:   interface.py
        new file:   shapefiles/scripts/directoryparse.py
        modified:   shapefiles/scripts/shapefile-generator.py

analyse.py has the directory loading code extracted into
directoryparse.py and the metadata construction extracted to
interface.py so only workflow definitions remain. It implicitly requires
dicts validate_deims_sites and validated_zones still.

app.R sources two new python files (sourcing many to avoid inter-module
imports and enforce decoupling) and uses a new interface-specific
function for creating DEIMS sites.

interface.py now contains all the python glue code required for the
shiny app to use all the modules successfully. When using the workflows
outside of the interface this file can be used as a reference for the
kind of work required.

directoryparse.py describes how to parse the shapefiles/ directory into
useful data and separates repo-specific logic from the shapefile
generation logic which others may want to use. Helper functions are
provided for e.g. loading/saving individual directories but most of the
time the loadAllInfo wrapper will be most helpful, which parses the
directory structure of the repo and returns it as dictionaries as
expected by the workflows.

shapefile-generator.py now works directly with geopandas objects so the
workflows can function without being tied to the repo's local file
structures. It also alters the data structures of deims sites and zones,
which has knock-on effects where this data is used. One exception is
generateMissingComposites which still relies on assumptions about
directory structure.

Full documentation of these changes is already underway and will be
completed in the new year.

The following changes also apply:
- composites directories are flattened to one level, each subdirectory
  containing only a composite .shp. Nuts names are changed to
  nutsYEAR-LEVEL to accommodate this
- removed raw directory and put boundaries at root of deims directories
- added lau2020 zones and composites for all sites
- restructure shapefiles/zones to three categories: lau, nuts (to
  prepare adding other nuts years) and national (to prepare adding other
  national zones)
- start replacing .format strings with f-strings and following PEP8