Enable multiselect of Fluxnet variables and nodata
In app.R, other than setting multiple = TRUE on wf3_variable, we also
replace it in the output filename with a timestamp, as it doesn't work
with a multiselect.

We also remove the now redundant global fluxnet_data read on startup and
add interpreting -9999 as nodata values for more accurate plots. As a
side-effect, the resultant values are converted to "NA" in output.

analyse.py is updated to accept multiple input variables, including by
truncating the plot title for simplicity's sake.