Automate _QC variables
We remove all _QC variables from the options in app.R and instead check
each variable for a corresponding _QC variable in analyse.py. This only
applies to the output, so _QC variables will not be plotted.

We also remove "RH" and rename "P" => "P_F" and "WS" => "WS_F" in the
variable choices.