wf3: Add provenance/acknowledgement metadata
Instead of just providing a csv of data to download, now a zip which
also includes a metadata file is provided.

To avoid writing metadata for data which is never downloaded, we provide
a separate function in `analyse.py` to write it instead - a strategy
which might eventually be applied to plotting previews and other
workflows too. The code_id_mapping dictionary is hardcoded for now (like
the options in the interface are anyway) but may eventually be
integrated into the site data shapefile scripts.

In `app.R` the site choices are updated to just represent the site
codes, which allows for easier processing in the metadata function. This
requires adjusting the wf3_filename variable. The save_wf3_output flow
is also rewritten to write the data and metadata to /tmp/ before zipping
them to the output folder as before.

The `input/fluxnet` and `input/icos` directories must now also contain
an `acknowledgement.txt` file and contain data files of the form
`{{sitecode}}.csv`.