Provide wf3 data in eLTER data format
To transform from the column-based data storage format provided in both
wf3 datasets, `wf3-data-prep.R` and `convertToElterFormat` are added.
The former should be run once when the wf3 data is first populated in
`input/fluxnet` and `input/icos`, the latter is run on every user
request to avoid rewriting `filterColumns` too extensively.

The mappings variable => unit and station code => deims ID are added as
constants to use in `convertToElterFormat`. To avoid duplication, the
latter is removed from `writeFilterColumnsMetadata` and an ID is passed
in instead.

Since all variables and their respective _QC columns are guaranteed to
exist in each dataset, the section on creating output_variables in
`filterColumns` is simplified, especially since the order of columns no
longer matters as they will be transformed in `app.R` anyway. The now
redundant `deims_site` parameter is also removed.

The TIME column is left as a character and a temporary column
PARSED_TIME is used for plotting to ensure that the ISO timestamp format
is preserved and also to avoid the bug mentioned in line 501.

The packrat environment is updated to include the tidyverse packages
used in `convertToElterFormat`.